requirements.txt
streamlit>=1.35.0
ultralytics>=8.3.0
openai>=1.30.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.26.0